requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap5==2023.10
Django==4.2.7
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-tinymce==3.6.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
